Move ModalDuplicateStrategy to modal folder and getUndercutStrategy to utils
